annotator/requirements.txt
python-decouple
fastapi
uvicorn[standard]
openai
pydantic